Cleaning up the Window class; You don't need duplicate entries for '(On)Paint' methods.
Also, I renamed isInside() to IsIn()- It's short, and to the point.

P.S. Jeff, when are you/we planning on splitting this off into individual client/server versions? Because
our planned "Desktop" component needs to be a "Server" object,  since our users don't need to mess around
with it's internal bits, unless you're overhauling or hacking(!) the UI environment.